Move the `by_region` method from Domain to Domains
The `Domains.by_region` syntax makes more sense than `Domain.by_region`.

It also returns a `Domains` array object that can have additional helper
methods defined onto it such as `find_by_name`
